requirement.txt
streamlit
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
